requirements.txt.txt
python-telegram-bot==20.7
python-docx==0.8.11